GAUD-8851: proposed changes
